[Bronze V] Title: 이상한 기호, Time: 80 ms, Memory: 11568 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -9,6 +9,6 @@ public class Main {
         StringTokenizer st = new StringTokenizer(br.readLine());
         int A = Integer.parseInt(st.nextToken());
         int B = Integer.parseInt(st.nextToken());
-        System.out.println(A * A - B * B);
+        System.out.println((long)A * A - (long)B * B);
     }
 }
